+++ app/src/androidTest/java/com/amoseui/cameraspecs/ExampleInstrumentedTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.amoseui.camerainfo
+package com.amoseui.cameraspecs
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -33,6 +33,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.amoseui.camerainfo", appContext.packageName)
+        assertEquals("com.amoseui.cameraspecs", appContext.packageName)
     }
 }
