+++ app/src/main/java/com/amoseui/cameraspecs/CameraIdsScreen.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.amoseui.camerainfo
+package com.amoseui.cameraspecs
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.material3.Text
