+++ app/src/main/java/com/amoseui/cameraspecs/CameraIdsViewModel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.amoseui.camerainfo
+package com.amoseui.cameraspecs
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
