+++ app/src/main/java/com/amoseui/cameraspecs/CameraInfoApplication.kt
@@ -14,14 +14,16 @@
  * limitations under the License.
  */
 
-package com.amoseui.camerainfo
+package com.amoseui.cameraspecs
 
-import kotlinx.coroutines.flow.Flow
-import javax.inject.Inject
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
 
-class CameraIdsRepository @Inject constructor(private val dataSource: CameraIdsSystemDataSource) {
-
-    val cameraIdsStream: Flow<List<CameraData.Camera>> = dataSource.cameraIdsStream
-
-    suspend fun refreshCameraIds() = dataSource.refreshCameraIds()
+@HiltAndroidApp
+class CameraInfoApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        Timber.plant(Timber.DebugTree())
+    }
 }
