+++ app/src/main/java/com/amoseui/cameraspecs/CameraResource.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.amoseui.camerainfo
+package com.amoseui.cameraspecs
 
 enum class CameraType(val value: Int) {
     NORMAL(0),
