+++ app/src/main/proto/camera.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option java_package = "com.amoseui.camerainfo";
+option java_package = "com.amoseui.cameraspecs";
 option java_multiple_files = true;
 
 message CameraData {
